Fix feedback form button into modal instead of page
Also remove hidden give feedback button on the main home page
